allow captures of up to 48MHz, if bus conditions permit
We allow the FX2 Waveform to operate using only a single state
(and thus at a rate equal to the max IFCLK of 48MHz) by allowing
the decision point state to re-execute its operation.

We accordingly remove the guards preventing operation with
zero-duration wait states; allowing waveforms with no wait states
to be generated.